Cache anchor positions until the buffer is edited
This avoids unnecessary lookups in the buffer tree when the anchor's
position has already been retrieved since the last edit.

We are representing the cache as a RefCell to avoid the need to acquire
a mutable reference to the buffer even if we just need to query it. If
the cache is already borrowed for some reason, we will simply skip
writing to it.